Added special case of 'install' for builder script

+++ build/hi3519a.sh
@@ -8,7 +8,10 @@ cd /home/zzlee/dev/qcap-dev/qcap
 mkdir build-$TARGET
 cd build-$TARGET
 
-rm lib/* -r
-rm bin/* -r
+if [ "$1" = "install" ]; then
+	rm lib/* -r
+	rm bin/* -r
+fi
+
 HI3519A_SDK_HOME=/home/zzlee/dev/hi3519a-sdk/Hi3519AV100_SDK_V2.0.2.0 cmake .. -DBUILD_TARGET=$TARGET
 make $@ -j 4
