+++ build/ubuntu1804_x64.sh
@@ -5,10 +5,13 @@
 TARGET=ubuntu1804_x64
 
 cd /home/zzlee/dev/qcap-dev/qcap
-mkdir build-$TARGET
+mkdir -p build-$TARGET
 cd build-$TARGET
 
-rm lib/* -r
-rm bin/* -r
+if [ "$1" = "install" ]; then
+	rm lib/* -r
+	rm bin/* -r
+fi
+
 cmake .. -DBUILD_TARGET=$TARGET
 make $@ -j 4
